Custom Class and Level Up Designer Tool v1
Added code for my Developer Tool. Opens a window that allows a developer to name a class, set stats for health, attack, defense, and speed, and set chance percentages each stat will increase on level up. When data is created, a new window pops up allowing the developer to see a test of how the stats increase for the class when it levels up.