Custom Class and Level Up Designer Tool v4
Added base stats for class data at level 1 called original stats. This stores what was the initial stats and remembers them. So when the designer is testing out how leveling up and stats work for the created class, they can click the return to base stats button and it will show what the original stats of the class are and the designer can test the level up system from the base stats and level 1 again.